+++ FinalProject.Api/Controllers/ControlController.cs
@@ -1,20 +1,27 @@
 ﻿using FinalProject.Business.Abstract;
 using FinalProject.Business.ValidationRules.FluentValidation;
+using FinalProject.Core.Utilities.Security.JWT;
 using FinalProject.Entities.DTOs;
 using FluentValidation.Results;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Linq;
 
 namespace FinalProject.Api.Controllers
 {
     [Route("[controller]")]
     [ApiController]
-    [Authorize(Roles = "Doctor")]
     public class ControlController : ControllerBase
     {
         private readonly IControlService _controlService;
-        public ControlController(IControlService controlService) => _controlService = controlService;
+        private readonly IJwtService _jwtService;
+
+        public ControlController(IControlService controlService, IJwtService jwtService)
+        {
+            _controlService = controlService;
+            _jwtService = jwtService;
+        }
 
         [HttpPost]
         public IActionResult Add(ControlDto control)
@@ -24,6 +31,10 @@ namespace FinalProject.Api.Controllers
             if (!result.IsValid)
                 foreach (var e in result.Errors)
                     return StatusCode(StatusCodes.Status400BadRequest, e.ErrorMessage);
+            var token = Request.Cookies["token"];
+            var verifyToken = _jwtService.VerifyToken(token);
+            var role = verifyToken.Data.Claims.FirstOrDefault(x => x.Type == RoleType.Type).Value;
+            if (!verifyToken.Success || role != "Doctor") return Unauthorized(new { message = "Bu əməliyyat üçün icazəniz yoxdur" });
             var resultService = _controlService.Add(control);
             if (!resultService.Success) return StatusCode(StatusCodes.Status400BadRequest, resultService.Message);
             return Ok(resultService);
@@ -37,6 +48,10 @@ namespace FinalProject.Api.Controllers
             if (!result.IsValid)
                 foreach (var e in result.Errors)
                     return StatusCode(StatusCodes.Status400BadRequest, e.ErrorMessage);
+            var token = Request.Cookies["token"];
+            var verifyToken = _jwtService.VerifyToken(token);
+            var role = verifyToken.Data.Claims.FirstOrDefault(x => x.Type == RoleType.Type).Value;
+            if (!verifyToken.Success || role != "Doctor") return Unauthorized(new { message = "Bu əməliyyat üçün icazəniz yoxdur" });
             var resultService = _controlService.Update(control);
             if (!resultService.Success) return StatusCode(StatusCodes.Status400BadRequest, resultService.Message);
             return Ok(resultService);
@@ -45,6 +60,10 @@ namespace FinalProject.Api.Controllers
         [HttpDelete]
         public IActionResult Delete(int id)
         {
+            var token = Request.Cookies["token"];
+            var verifyToken = _jwtService.VerifyToken(token);
+            var role = verifyToken.Data.Claims.FirstOrDefault(x => x.Type == RoleType.Type).Value;
+            if (!verifyToken.Success || role != "Doctor") return Unauthorized(new { message = "Bu əməliyyat üçün icazəniz yoxdur" });
             var result = _controlService.Delete(id);
             if (result.Success) return Ok(result);
             return BadRequest(result);
@@ -53,6 +72,10 @@ namespace FinalProject.Api.Controllers
         [HttpGet("all")]
         public IActionResult GetAll()
         {
+            var token = Request.Cookies["token"];
+            var verifyToken = _jwtService.VerifyToken(token);
+            var role = verifyToken.Data.Claims.FirstOrDefault(x => x.Type == RoleType.Type).Value;
+            if (!verifyToken.Success || role != "Doctor") return Unauthorized(new { message = "Bu əməliyyat üçün icazəniz yoxdur" });
             var result = _controlService.GetAll();
             if (result.Success) return Ok(result);
             return BadRequest(result);
